package updates and slight refactoring (#145)
