+++ app/main.dev.js
@@ -167,8 +167,6 @@ app.on('ready', async () => {
     fullscreenable: false,
     backgroundColor: '#2f8da3',
     webPreferences: {
-      // Prevents renderer process code from not running when window is
-      // hidden
       backgroundThrottling: false,
       nodeIntegration: true,
       enableRemoteModule: true
@@ -198,11 +196,6 @@ app.on('ready', async () => {
 
   toolbarWindow.loadURL(`file:///${__dirname}/${appPath}?toolbar`);
 
-  // @TODO: Use 'ready-to-show' event
-  //        https://github.com/electron/electron/blob/master/docs/api/browser-window.md#using-ready-to-show-event
-  //
-  //
-
   toolbarWindow.webContents.on('did-finish-load', () => {
     if (!toolbarWindow) {
       throw new Error('"toolbarWindow" is not defined');
@@ -218,7 +211,4 @@ app.on('ready', async () => {
 
   const menuBuilder = new MenuBuilder(toolbarWindow);
   menuBuilder.buildMenu();
-
-  // Remove this if your app does not use auto updates
-  // eslint-disable-next-line
 });
