+++ app/toolbar/components/common/header/Header.scss
