+++ app/toolbar/components/common/popup/Popup.scss
