+++ app/toolbar/components/status/states/confirm/Confirm.scss
