+++ app/toolbar/components/status/states/error/Error.scss
