+++ app/toolbar/components/status/states/running/Running.scss
