+++ app/toolbar/components/status/states/stopped/Stopped.scss
