+++ app/toolbar/components/status/states/transition/Transition.scss
