+++ package.json
@@ -290,7 +290,7 @@
     "keytar": "^6.0.1",
     "marked": "^0.3.19",
     "moment": "^2.24.0",
-    "node-fetch": "^2.5.0",
+    "node-fetch": "^2.6.1",
     "node-gyp": "^7.1.0",
     "node-powershell": "^4.0.0",
     "normalize-sass": "^1.0.0",
